Crate target_spec

Evaluate Cargo.toml target specifications against platform triples.

Cargo supports platform-specific dependencies. These dependencies can be specified in one of two ways:

# 1. As Rust-like `#[cfg]` syntax.
[target.'cfg(all(unix, target_arch = "x86_64"))'.dependencies]
native = { path = "native/x86_64" }

# 2. Listing out the full target triple.
[target.x86_64-pc-windows-gnu.dependencies]
winhttp = "0.4.0"

target-spec provides the eval API which can be used to figure out whether such a dependency will be included on a particular platform.

use target_spec::eval;

// Evaluate Rust-like `#[cfg]` syntax.
let cfg_target = "cfg(all(unix, target_arch = \"x86_64\"))";
assert_eq!(eval(cfg_target, "x86_64-unknown-linux-gnu"), Ok(Some(true)));
assert_eq!(eval(cfg_target, "i686-unknown-linux-gnu"), Ok(Some(false)));
assert_eq!(eval(cfg_target, "x86_64-pc-windows-msvc"), Ok(Some(false)));

// Evaluate a full target-triple.
assert_eq!(eval("x86_64-unknown-linux-gnu", "x86_64-unknown-linux-gnu"), Ok(Some(true)));
assert_eq!(eval("x86_64-unknown-linux-gnu", "x86_64-pc-windows-msvc"), Ok(Some(false)));

For more advanced usage, see Platform and TargetSpec.

Optional features

summaries: Adds the summaries module to enable serialization of Platform and TargetFeatures.

proptest1: Enables support for property-based testing of Platform and TargetFeatures using proptest.

Minimum supported Rust version

The minimum supported Rust version (MSRV) is:

For target-spec 1.0.x: Rust 1.54.

For target-spec 1.1.x: Rust 1.56.

For target-spec 1.2.x: Rust 1.58.

For target-spec 1.3.x: Rust 1.62.

Within the 1.x series, MSRV bumps will be accompanied by a minor version update.

Related crates

To pretty-print target-spec errors, consider using the miette diagnostic library with target-spec-miette.

Re-exports

pub use errors::Error;

Modules

errors

Errors returned by target-spec.

summariessummaries

Serialized versions of platform and target features.

Structs

Platform

A platform to evaluate target specs against.

TargetExpression

A target expression.

Triple

A single, specific target, uniquely identified by a triple.

Enums

TargetFeatures

A set of target features to match.

TargetSpec

A parsed target specification or triple, as found in a Cargo.toml file.

Functions

eval

Evaluates the given spec against the provided target and returns Some(true) on a successful match, and Some(false) on a failing match.