Module target_spec::custom_platforms

Support for creating custom platforms.

This module re-exports parts of the cfg_expr dependency to allow creating platforms unknown to rustc by default.

Structs

Arch
The "architecture" field

Env
The "environment" field, which specifies an ABI environment on top of the operating system. In many configurations, this field is omitted, and the environment is implied by the operating system.

Os
The "operating system" field, which sometimes implies an environment, and sometimes isn't an actual operating system.

TargetInfo
Contains information regarding a particular target known to rustc

Vendor
The "vendor" field, which in practice is little more than an arbitrary modifier.

Enums

Endian
The endian types known to rustc

Family
All of the target families known to rustc