Module target_spec::custom_platforms::cfg_expr

Parts of the cfg_expr dependency.

Structs

Arch
The “architecture” field

Env
The “environment” field, which specifies an ABI environment on top of the operating system. In many configurations, this field is omitted, and the environment is implied by the operating system.

Family
The target family, which describes a set of targets grouped in some logical manner, typically by operating system. This includes values like unix and windows.

Os
The “operating system” field, which sometimes implies an environment, and sometimes isn’t an actual operating system.

TargetInfo
Contains information regarding a particular target known to rustc

Triple
The unique identifier for a target.

Vendor
The “vendor” field, which in practice is little more than an arbitrary modifier.

Enums

Endian
The endian types known to rustc