Function target_spec::eval

pub fn eval(spec_or_triple: &str, platform: &str) -> Result<Option<bool>, Error>

Evaluates the given spec against the provided target and returns Some(true) on a successful match, and Some(false) on a failing match.

This defaults to treating target features as unknown, and returns None if the overall result is unknown.

For more advanced uses, see TargetSpec::eval.

For more information, see the crate-level documentation.