Struct target_spec::Platform

pub struct Platform { /* private fields */ }

A platform to evaluate target specs against.

Implementations

impl Platform

pub fn new(
triple_str: impl Into<Cow<'static, str>>,
target_features: TargetFeatures
) -> Result<Self, Error>

Creates a new Platform from the given triple and target features.

Returns an error if this platform wasn’t known to target-spec.

pub fn current() -> Result<Self, Error>

Returns the current platform, as detected at build time.

This will return an error if the current platform was unknown to this version of target-spec.

pub fn from_triple(triple: Triple, target_features: TargetFeatures) -> Self

Creates a new platform from a Triple and target features.

pub fn add_flags(
&mut self,
flags: impl IntoIterator<Item = impl Into<Cow<'static, str>>>
)

Adds a set of flags to accept.

A flag is a single token like the foo in cfg(not(foo)).

A default cargo build will always evaluate flags to false, but custom wrappers may cause some flags to evaluate to true. For example, as of version 0.6, cargo web build will cause cargo_web to evaluate to true.

pub fn triple_str(&self) -> &str

Returns the target triple string for this platform.

pub fn flags(&self) -> impl Iterator<Item = &str> + ExactSizeIterator

Returns the set of flags enabled for this platform.

pub fn has_flag(&self, flag: impl AsRef<str>) -> bool

Returns true if this flag was set with add_flags.

pub fn triple(&self) -> &Triple

Returns the underlying Triple.

pub fn target_features(&self) -> &TargetFeatures

Returns the set of target features for this platform.

impl Platform

Helpers for property testing

The methods in this section allow Platform instances to be used in property-based testing scenarios.

Currently, proptest 1 is supported if the proptest1 feature is enabled.

pub fn strategy(
target_features: impl Strategy<Value = TargetFeatures>
) -> impl Strategy<Value = Platform>

This is supported on crate feature proptest1 only.

Given a way to generate TargetFeatures instances, this returns a Strategy that generates a platform at random.

Requires the proptest1 feature to be enabled.

Examples

use proptest::prelude::*;
use target_spec::{Platform, TargetFeatures};

// target_features is a strategy that always produces TargetFeatures::Unknown.
let target_features = Just(TargetFeatures::Unknown);
let strategy = Platform::strategy(target_features);

pub fn filtered_strategy(
triple_filter: impl Fn(&'static str) -> bool,
target_features: impl Strategy<Value = TargetFeatures>
) -> impl Strategy<Value = Platform>

This is supported on crate feature proptest1 only.

A version of strategy that allows target triples to be filtered.

Requires the proptest1 feature to be enabled.

impl Platform

pub fn to_summary(&self) -> PlatformSummary

This is supported on crate feature summaries only.

Converts this Platform to a serializable form.

Requires the summaries feature to be enabled.